add complexn3.
